chore: refactor edge node empty state to use ListTableBlock emptyBlock prop
